Finish migrate container
