+++ python/tvm/api.py
@@ -19,8 +19,11 @@
 from numbers import Integral as _Integral
 
 import tvm._ffi
+import tvm.ir
 
 from tvm.runtime import convert, const, DataType
+from tvm.ir import container as _container
+
 from ._ffi.base import string_types, TVMError
 from ._ffi.registry import register_func, get_global_func, extract_ext_funcs
 
@@ -29,7 +32,6 @@ from . import make as _make
 from . import expr as _expr
 from . import tensor as _tensor
 from . import schedule as _schedule
-from . import container as _container
 from . import tag as _tag
 
 int8 = "int8"
@@ -68,29 +70,6 @@ def max_value(dtype):
         The maximum value of dtype.
     """
     return _api_internal._max_value(dtype)
-
-
-def get_env_func(name):
-    """Get an EnvFunc by a global name.
-
-    Parameters
-    ----------
-    name: str
-        The name of the global function.
-
-    Returns
-    -------
-    env_func : EnvFunc
-        The result env function.
-
-    Note
-    ----
-    EnvFunc is a Object wrapper around
-    global function that can be serialized via its name.
-    This can be used to serialize function field in the language.
-    """
-    return _api_internal._EnvFuncGet(name)
-
 
 def var(name="tindex", dtype=int32):
     """Create a new variable with specified name and dtype
@@ -649,7 +628,7 @@ def _IterVar(dom, name, iter_type, thread_tag=''):
                 raise TypeError("need to be list of ranges")
             dom = Range(dom[0], dom[1])
 
-        if not isinstance(dom, _container.Range):
+        if not isinstance(dom, tvm.ir.Range):
             raise TypeError("dom need to be Range")
     name = name if name else 'iter'
     v = var(name)
