+++ python/tvm/ir/base.py
@@ -17,8 +17,10 @@
 """Common base structures."""
 import tvm._ffi
 
-from tvm.runtime import Object
+import tvm.error
 import tvm.runtime._ffi_node_api
+from tvm.runtime import Object
+
 from . import _ffi_api
 from . import json_compact
 
@@ -87,6 +89,31 @@ class Span(Object):
             _ffi_api.Span, source, lineno, col_offset)
 
 
+@tvm._ffi.register_object
+class EnvFunc(Object):
+    """Environment function.
+
+    This is a global function object that can be serialized by its name.
+    """
+    def __call__(self, *args):
+        return _ffi_api.EnvFuncCall(self, *args)
+
+    @property
+    def func(self):
+        return _ffi_api.EnvFuncGetPackedFunc(self)
+
+    @staticmethod
+    def get(name):
+        """Get a static env function
+
+        Parameters
+        ----------
+        name : str
+            The name of the function.
+        """
+        return _ffi_api.EnvFuncGet(name)
+
+
 def load_json(json_str):
     """Load tvm object from json_str.
 
@@ -103,7 +130,7 @@ def load_json(json_str):
 
     try:
         return tvm.runtime._ffi_node_api.LoadJSON(json_str)
-    except TVMError:
+    except tvm.error.TVMError:
         json_str = json_compact.upgrade_json(json_str)
         return tvm.runtime._ffi_node_api.LoadJSON(json_str)
