+++ tests/python/relay/test_type_functor.py
@@ -64,7 +64,7 @@ def test_tuple_type():
 
 
 def test_type_relation():
-    func = tvm.get_env_func('tvm.relay.type_relation.Broadcast')
+    func = tvm.ir.EnvFunc.get('tvm.relay.type_relation.Broadcast')
     attrs = tvm.make.node('attrs.TestAttrs', name='attr', padding=(3,4))
     tp = TypeVar('tp')
     tf = FuncType([], TupleType([]), [], [])
